Changed atom_state for atoms_state
