+++ simulationdataschema/model_system.py
@@ -304,7 +304,7 @@ class AtomicCell(Cell):
     A base section used to specify the atomic cell information of a system.
     """
 
-    atoms_state = SubSection(sub_section=AtomState.m_def, repeats=True)
+    atoms_state = SubSection(sub_section=AtomsState.m_def, repeats=True)
 
     equivalent_atoms = Quantity(
         type=np.int32,
@@ -521,7 +521,7 @@ class Symmetry(ArchiveSection):
         atomic_cell = AtomicCell(type=cell_type)
         atomic_cell.lattice_vectors = cell * ureg.angstrom
         for label, atomic_number in zip(labels, atomic_numbers):
-            atom_state = AtomState(chemical_symbol=label, atomic_number=atomic_number)
+            atom_state = AtomsState(chemical_symbol=label, atomic_number=atomic_number)
             atomic_cell.m_add_sub_section(AtomicCell.atoms_state, atom_state)
         atomic_cell.positions = (
             positions * ureg.angstrom
